signatures and auth abstractions

+++ App/Application/Services/todos.service.ts
@@ -1,30 +1,63 @@
 import TodoEntity from '../../Domain/Todo/todo.entity';
+import TodoRepositoryI from '../../Domain/Todo/todo.repository';
 import PaginationOptions from '../../Domain/Utils/Pagination/pagination.options';
-import SequelizeTodoRepository from '../../Infrastructure/MySqlrepository/todo.repository';
+import CustomError from '../../Infrastructure/Exceptions/custom-error';
 
 class TodosService {
-    public repository
-    constructor(sequelizeTodoRepository: SequelizeTodoRepository){
-        this.repository = sequelizeTodoRepository
+  public repository;
+
+  constructor(todoRepository: TodoRepositoryI) {
+    this.repository = todoRepository;
+  }
+
+  public async getTodos(userId: string, pagination: PaginationOptions) {
+    const result = await this.repository.fetchAll(userId, pagination);
+
+    return result.getPaginatedData();
+  }
+
+  public async createTodo(userId: string, name: string) {
+    if (!name) {
+      throw new CustomError(400, 'Must provide a name');
     }
-    public async getTodos(pagination:PaginationOptions){
-        return this.repository.fetchAll(pagination)
+
+    const todoEntity = TodoEntity.createFromInput(userId, name);
+
+    await this.repository.addItem(todoEntity);
+    return;
+  }
+
+  public async updateTodo(id: string, name: string) {
+    const todoEntity = await this.repository.fetchById(id);
+
+    if (!todoEntity) {
+      throw new CustomError(400, 'Resource not found');
     }
-    public async createTodo(name:string) {
-        const todoEntity = TodoEntity.createFromInput(name)
-        return this.repository.addItem(todoEntity)
+
+    todoEntity.name = name;
+
+    return this.repository.editItem(todoEntity);
+  }
+
+  public async deleteTodo(id: string) {
+    const todoEntity = await this.repository.fetchById(id);
+
+    if (!todoEntity) {
+      throw new CustomError(400, 'Resource not found');
     }
-    public async updateTodo(id:string, body:any){
-        const todoEntity:TodoEntity = await this.getTodo(id);
-        todoEntity.name = body.name
-        return this.repository.editItem(todoEntity)
+
+    return this.repository.removeItem(todoEntity);
+  }
+
+  public async getTodo(id: string) {
+    const todoEntity = await this.repository.fetchById(id);
+
+    if (!todoEntity) {
+      throw new CustomError(400, 'Resource not found');
     }
-    public async deleteTodo(id:string){
-        const todoEntity:TodoEntity = await this.getTodo(id);
-        return this.repository.removeItem(todoEntity)
-    }
-    public async getTodo(id:string){
-        return this.repository.fetchById(id)
-    }
+
+    return todoEntity;
+  }
 }
-export default TodosService
+
+export default TodosService;
